+++ src/common/Crypt/base64.cpp
@@ -89,7 +89,7 @@ std::string Sapphire::Common::Util::base64Encode( uint8_t const* bytes_to_encode
 
 std::string Sapphire::Common::Util::base64Decode( std::string const& encoded_string )
 {
-  int32_t in_len = encoded_string.size();
+  int32_t in_len = (int32_t)encoded_string.size();
   int32_t i = 0;
   int32_t j = 0;
   int32_t in_ = 0;
